+++ processor/src/parallel/call.rs
@@ -5,9 +5,9 @@ use miden_core::{
     mast::{CallNode, MastNodeExt},
 };
 
-use super::{CoreTraceFragmentGenerator, trace_builder::OperationTraceConfig};
+use super::{CoreTraceFragmentFiller, trace_builder::OperationTraceConfig};
 
-impl CoreTraceFragmentGenerator {
+impl<'a> CoreTraceFragmentFiller<'a> {
     /// Adds a trace row for the start of a CALL/SYSCALL operation.
     pub fn add_call_start_trace_row(
         &mut self,