+++ processor/src/parallel/split.rs
@@ -6,9 +6,9 @@ use miden_core::{
     mast::{MastNodeExt, SplitNode},
 };
 
-use super::{CoreTraceFragmentGenerator, trace_builder::OperationTraceConfig};
+use super::{CoreTraceFragmentFiller, trace_builder::OperationTraceConfig};
 
-impl CoreTraceFragmentGenerator {
+impl<'a> CoreTraceFragmentFiller<'a> {
     /// Adds a trace row for the start of a SPLIT operation.
     ///
     /// This is a convenience method that calls `add_split_trace_row` with `TraceRowType::Start`.